log: move LogPanel to log package
